Worked on adding game states and fixing bug
it now has an object to hold the states of the game and theoretically tracks the state of the game. After voting it also should add to the players scores from the votes on the cards. Also I made it so the voted cards are stored in objects based on the rooms they're in in the server.